requirements.txt
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
Flask==2.3.2
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
pluggy==1.0.0
pybase64==1.2.3
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.30.0
six==1.16.0
tomli==2.0.1
tzdata==2023.3
urllib3==2.0.2
Werkzeug==2.3.4
